Add DQT v3 API client
